Moving shared methods in spec helper
